+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version 10.2.0
 ## New
 - **allgemein**: Sicherheitsupdates eingespielt.
+- **lux-autocomplete**: Property "luxFilterFn" hinzugefügt. [Issue 40](https://github.com/IHK-GfI/lux-components/issues/40)
 
 # Version 10.1.0
 ## New